+++ config/arbitrum.ts
@@ -126,7 +126,11 @@ export default <NetworkData>{
                 },
             },
         },
-        stakewise: '0xf7d4e7273e5015c96728a6b02f31c505ee184603',
+        stakewise: {
+            url: 'https://mainnet-graph.stakewise.io/subgraphs/name/stakewise/stakewise',
+            token: '0xf7d4e7273e5015c96728a6b02f31c505ee184603',
+        },
+        etherfi: '0x35751007a407ca6feffe80b3cb397736d2cf4dbe',
         defaultHandlers: {
             wstETH: {
                 tokenAddress: '0x5979d7b546e38e414f7e9822514be443a4800529',