+++ config/gnosis.ts
@@ -17,7 +17,7 @@ export default <NetworkData>{
         blocks: 'https://api.studio.thegraph.com/query/48427/gnosis-blocks/version/latest',
         gauge: `https://gateway-arbitrum.network.thegraph.com/api/${env.THEGRAPH_API_KEY_BALANCER}/deployments/id/Qme9hQY1NZ8ReVDSSQb893s2fGpeLkgfwXd3YU5rndACaP`,
         aura: 'https://data.aura.finance/graphql',
-        cowAmm: `https://gateway-arbitrum.network.thegraph.com/api/${env.THEGRAPH_API_KEY_BALANCER}/deployments/id/QmP37o4UgJZVEDxrwx2V1hkCZACLFaMfS6r5yzcK13wHeS`,
+        cowAmm: `https://gateway-arbitrum.network.thegraph.com/api/${env.THEGRAPH_API_KEY_BALANCER}/deployments/id/QmVAZTFvvkCkkBvXQ2mrJfnW349UVSHJdgwA1u7QYau6iw`,
     },
     eth: {
         address: '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee',
@@ -34,12 +34,9 @@ export default <NetworkData>{
         platformId: 'xdai',
         excludedTokenAddresses: [],
     },
-    rpcUrl:
-        env.GATEWAYFM_API_KEY && (env.DEPLOYMENT_ENV as DeploymentEnv) === 'main'
-            ? `https://rpc.eu-central-2.gateway.fm/v4/gnosis/archival/mainnet?apiKey=${env.GATEWAYFM_API_KEY}`
-            : env.GROVE_CITY
-            ? `https://gnosischain-mainnet.rpc.grove.city/v1/${env.GROVE_CITY}`
-            : 'https://gnosis.drpc.org',
+    rpcUrl: env.ALCHEMY_API_KEY
+        ? `https://gnosis-mainnet.g.alchemy.com/v2/${env.ALCHEMY_API_KEY}`
+        : 'https://gnosis.drpc.org',
     rpcMaxBlockRange: 2000,
     protocolToken: 'bal',
     bal: {
@@ -68,11 +65,21 @@ export default <NetworkData>{
     multicall3: '0xca11bde05977b3631167028862be2a173976ca11',
     avgBlockSpeed: 1,
     ybAprConfig: {
+        stakewise: {
+            url: 'https://gnosis-graph.stakewise.io/subgraphs/name/stakewise/stakewise',
+            token: '0xf490c80aae5f2616d3e3bda2483e30c4cb21d1a0',
+        },
         defaultHandlers: {
             wstETH: {
                 tokenAddress: '0x6c76971f98945ae98dd7d4dfca8711ebea946ea6',
                 sourceUrl: 'https://eth-api.lido.fi/v1/protocol/steth/apr/sma',
                 path: 'data.smaApr',
+                isIbYield: true,
+            },
+            rETH: {
+                tokenAddress: '0xc791240d1f2def5938e2031364ff4ed887133c3d',
+                sourceUrl: 'https://rocketpool.net/api/mainnet/payload',
+                path: 'rethAPR',
                 isIbYield: true,
             },
         },