+++ modules/controllers/pools-debug.test.ts
@@ -32,6 +32,17 @@ describe('pool controller debugging', () => {
         await CowAmmController().updateSurplusAprs();
     }, 5000000);
 
+    it('cow snapshots', async () => {
+        await CowAmmController().addPools('1');
+        // await CowAmmController().addPools('1');
+        await CowAmmController().syncSwaps('1');
+        // await CowAmmController().syncSwaps('1');
+        await CowAmmController().syncJoinExits('1');
+        await CowAmmController().updateVolumeAndFees('1');
+        await CowAmmController().syncSnapshots('1')
+        await CowAmmController().updateSurplusAprs();
+    }, 5000000);
+
     describe('pool debugging', () => {
         it('reload pools', async () => {
             //only do once before starting to debug