+++ modules/network/apr-config-types.ts
@@ -13,7 +13,14 @@ export interface YbAprConfig {
     tetu?: TetuAprConfig;
     tranchess?: TranchessAprConfig;
     yearn?: YearnAprConfig;
-    stakewise?: string;
+    stakewise?: {
+        url: string;
+        token: string;
+    };
+    maple?: {
+        url: string;
+        token: string;
+    };
     etherfi?: string;
     sveth?: boolean;
     defaultHandlers?: DefaultHandlerAprConfig;